new test data - for binary TxContext
